[kibana] Fix typo on 'when' gate for index
The default mapping should be added if the 'defaultIndex' is already
configured in Kibana's "config" entry. This check used to check if this
config was set, which would always be true due to it being explicitly
set in the "Change kibana config to set index as defaultIndex" step. So,
it would always be skipped.