Support for specifying globals file location
Instead of assuming that the globals.yml file is in the config
directory, follow the same approach as for passwords, and allow the user
to specify explicitly. Defaults to $config_dir/globals.yml. This makes
the --configdir less important; if globals location is overridden the
configdir only really affects where the adminrc.sh file is dropped
during post-deploy.