[nova] Use keystone_admin_url for Ironic config
It's not necessarily the case that the openstack_auth object, used to
register/write objects to Keystone, is the same Keystone deployment that
Ironic should be configured to point to. In general, the Keystone server
being deployed should be used here (defined as keystone_admin_url). This
also matches how Keystone is used elsewhere in the configuration.

Change-Id: I930e263a687b8d0ce1fce1ae0e8d0704d70eb7c1