[haproxy-config] Allow disabling frontend/backend
In split mode, allow disabling either a backend or a frontend. This is
really only useful to allow one frontend to service multiple backends,
and that's only really useful when you want to do something like
host-based routing over a shared port (e.g., 443). We want to allow
Horizon's listen section to route to other backends depending on the
destination port, so we can host Keycloak as well as Horizon on 443.

Change-Id: If4b4698e8e1ae134957a8e93b51fbd0acece7109