Add kolla_base_distro_major_version flag
There are a few cases where the host/deploy OS version is assumed to
equal the container image OS major version. This breaks down when
trying to deploy CentOS 8 containers on CentOS 7. This adds a new
flag ``kolla_base_distro_major_version``, which can be used to override
the auto-detect of the container OS version from the host version.

E.g., it can be set like::

  kolla_base_distro_major_version: 8

For deploying CentOS 8 on CentOS 7.

Change-Id: I4d4863923254746fbecab4ba449969ffaa163b95