Force elasticsearch6 compatible config paths
Elasticsearch6 uses /etc/elasticsearch as the config path, and is included
in our ubuntu container builds. As part of the initial migration to es6,
a separate build target (elasticsearch6) was mad in kolla. However, this
was never implemented for ubuntu based containers.

Since we no longer deploy elasticsearch older than 6 anywhere, removing this
conditional from the config template allows es6 to function properly on ubuntu.

During the openstack version upgrade, we should drop other references to this.