Have HAProxy load its own certs at startup
Part of the Let's Encrypt changes tried to have certbot be fully in
charge of managing Let's Encrypt certs. This didn't really work at
startup, and the reload script had to be run manually. With this change,
HAProxy will load the Let's Encrypt certs by itself on startup, and
certbot will only be in charge of reloading it when the certs are
renewed live.