Allow to run post-deploy without privilege escalation
In deployments where `kolla-ansible` is run by a regular user,
`post-deploy` could fail when attempting to create the
`/etc/openstack` directory if privilege escalation is not
enabled/possible non-interactively.

With this change, we can run `post-deploy` with `become: false` if
root is not the owner of the `node_config` directory.

This change also removes the task creating `/etc/openstack` as this
directory is unused since the change moving the cloud config file
to `node_config`.

Related-Id: Idd458ad6ef29e4eee2f9e537b4eae39d26eb9f64
Closes-Bug: #2098887
Change-Id: Ied5ef4e1e3c9346342962d44de26702522b06b86

(cherry picked from commit 1cf04e8058391c34b5603bc1a380d1c6422c6f8f)