calico: always use worker NodeIP
due to our routing implementation, if the calico IP does not match the
kube node IP, the control plane services cannot reach endpoints with
host networking. Since we need both of them (calico and nodeip) to be
the wireguard interface, set it at the worker level, and just have
calico match.